LayoutTests/platform/wk2/http/tests/loading/basic-credentials-sent-automatically-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with first:first-pw
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame
frame "<!--frame2-->" - didStartProvisionalLoadForFrame
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Responding with second:second-pw
frame "<!--frame2-->" - didCommitLoadForFrame
frame "<!--frame2-->" - didFinishDocumentLoadForFrame
frame "<!--frame2-->" - didHandleOnloadEventsForFrame
frame "<!--frame2-->" - didFinishLoadForFrame
frame "<!--frame3-->" - didStartProvisionalLoadForFrame
frame "<!--frame3-->" - didCommitLoadForFrame
frame "<!--frame3-->" - didFinishDocumentLoadForFrame
frame "<!--frame3-->" - didHandleOnloadEventsForFrame
frame "<!--frame3-->" - didFinishLoadForFrame
frame "<!--frame4-->" - didStartProvisionalLoadForFrame
frame "<!--frame4-->" - didCommitLoadForFrame
frame "<!--frame4-->" - didFinishDocumentLoadForFrame
frame "<!--frame4-->" - didHandleOnloadEventsForFrame
frame "<!--frame4-->" - didFinishLoadForFrame
This test makes sure that once an HTTP Basic Auth. protected path is authenticated once, urls that emanate from that path automatically have their credentials sent without a challenge.
The first frame's path is /loading/resources/subresources/protected-resource.php, and we should get a challenge for it.
It will be authorized with first/first-pw.
The second frame's path is /loading/resources/protected-resource.php, and we should get a challenge for it, because it does not share a common subpath of the previously authorized resource.
It will be authorized with second/second-pw.
The third frame's path is also /loading/resources/protected-resource.php, and we should *not* get a challenge for it because we authorized to this path for the second frame.
It will be authorized with second/second-pw.
The fourth frame's path is /loading/resources/othersubresources/protected-resource.php, and we should *not* get a challenge for it, because it has a common subpath with the previously authorized second and third frames.
It will be authorized with second/second-pw.


--------
Frame: '<!--frame1-->'
--------
Authenticated as user: first password: first-pw

--------
Frame: '<!--frame2-->'
--------
Authenticated as user: second password: second-pw

--------
Frame: '<!--frame3-->'
--------
Authenticated as user: second password: second-pw

--------
Frame: '<!--frame4-->'
--------
Authenticated as user: second password: second-pw
